Enter payment is left

+++ pos/src/main/java/se/kth/iv1350/dto/ItemDTO.java
@@ -7,13 +7,13 @@ public class ItemDTO {
     public final int itemID;
     public final String name;
     public final int cost;
-    public final float vat;
+    public final double vat;
     public final String description;
 
     /**
      * Instantiates a new object of the ItemDTO type containing the information of an item.
      */
-    public ItemDTO(int itemID, String name, int cost, float vat, String description)
+    public ItemDTO(int itemID, String name, int cost, double vat, String description)
     {
         this.itemID = itemID;
         this.name = name;
